agents/travel_planner/requirements.txt
streamlit>=1.40.0
google-genai>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
folium>=0.18.0
streamlit-folium>=0.23.0
